Allow calling recursive functions at load time
Mark the refs in lams of decls as refs of the nearest app in the call
stack (if any) rather than refs of the current decl, since they just
need to have been defined when the app is evaluated rather than when the
current decl is loaded. This allows (mutually) recursive functions to be
called in apps in decls but still disallows cyclic decls.